+++ CHANGES.rst
@@ -8,6 +8,9 @@ Changelog of lizard-auth-server
 - Added ``/api2/organisations/`` that lists the organisation names plus
   unique ID.
 
+- Added more logging and made 'duplicate username' error, when creating a
+  user, more explicit.
+
 
 2.5 (2016-09-23)
 ----------------